API: Move state to top-level
