+++ pcdsdevices/epics/attenuator.py
@@ -13,7 +13,7 @@ from ophyd.status import wait as status_wait
 from ophyd import (Device, EpicsSignal, EpicsSignalRO, Component,
                    FormattedComponent)
 
-from .state import SubscriptionStatus
+from ..state import SubscriptionStatus
 
 logger = logging.getLogger(__name__)
 MAX_FILTERS = 12
